Links_share/requirements.txt
python>=3.0
python-telegram-bot==13.3
